Update to use latest Twitch minecraftinstance.json structure
So for now, updating will be broken, I don't have the time to look into getting it working and honestly with everything going on, it's just easier to go to the Twitch Launcher, update the mods, copy the manifest (or portions of the manifest) to their respective locations in the manifest(s) in the project's git repository.  For now, updating will be disabled along with merging manifests and MPD will simply crank through the Twitch manifest and download mods for pack development.

Yes, this broke our tests, because our tests were using the old structure.